Final test config with Ganache
